chore: add tests + analytics
