+++ lib/analytics/analytics_constants.dart
@@ -74,8 +74,6 @@ class AnalyticsScreenNames {
   static const savedSearchServiceCiviqueDelete = "/saved_search/service-civique/delete";
 
   static const suppressionAccount = "/profil/suppression-compte";
-  static const suppressionAccountConfirmation = "/profil/suppression-compte/confirm";
-  static const suppressionAccountSucceded = "/login?deleteSuccess=true";
 }
 
 class AnalyticsActionNames {
@@ -117,6 +115,9 @@ class AnalyticsActionNames {
   static String deleteSavedSearchAlternance = "/saved_searches/alternance/list?deleteSuccess=true";
   static String deleteSavedSearchImmersion = "/saved_searches/immersion/list?deleteSuccess=true";
   static String deleteSavedSearchServiceCivique = "/saved_searches/service_civique/list?deleteSuccess=true";
+
+  static const suppressionAccountConfirmation = "/profil/suppression-compte/confirm";
+  static const suppressionAccountSucceded = "/login?deleteSuccess=true";
 }
 
 class AnalyticsCustomDimensions {
